// app/models/event.tsx
export interface Event {
    id: string;
    signUpMethod: string;
    dayOfWeek: string;
    showTime: string;
    mapUrl: string;
    name: string;
    rrule: string;
    neighborhood: string;
    listMethod: string;
    recurrenceDescription: string;
    length: string;
    notes: string;
    host: string;
    isActive: boolean;
    mictype: string;
    signupTime: string;
  }